separate configparser into non-workflow part
